refactor(discussion): correct a problem in entities
